requirements.txt
mss
keyboard
pyuserinput
